Move @pollyjs/persister type overrides into `types/` folder.
